// src/models/enum-models/TaskManagerEnum.ts
export enum TaskMangerStatusEnum {
  CREATED = "Created",
  NEW = "New",
  TO_DO = "To Do",
  IN_PROGRESS = "In Progress",
  BLOCKED = "Blocked",
  ON_HOLD = "On Hold",
  UNDER_REVIEW = "Under Review",
  ARCHIVED = "Archived",
  CANCELLED = "Cancelled",
  DELETED = "Deleted",
  COMPLETED = "Completed",
  DONE = "Done",
}

export enum TaskMangerTypeEnum {
  ALL = "ALL",
  SOCIAL_MEDIA_POST = "SOCIAL_MEDIA_POST",
}